rename: groupped common components
