+++ src/components/common/index.ts
@@ -1,4 +1,4 @@
-import Button from '@/components/button';
+import Button from '@/components/common/button';
 import CustomLink from './customLink';
 import Divider from './divider';
 import InputField from './inputField';
