+++ src/pages/create-account.tsx
@@ -1,4 +1,4 @@
-import { Button, CustomLink, Divider, InputField } from '@/components';
+import { Button, CustomLink, Divider, InputField } from '@/components/common';
 import useFetch from '@/libs/client/useFetch';
 import { Join } from '@/types/auth';
 import { useRouter } from 'next/router';
